+++ wp-content/plugins/newspack-popups/api/campaigns/class-maybe-show-campaign.php
@@ -19,8 +19,6 @@ class Maybe_Show_Campaign extends Lightweight_API {
 
 	/**
 	 * Constructor.
-	 *
-	 * @codeCoverageIgnore
 	 */
 	public function __construct() {
 		parent::__construct();
@@ -29,17 +27,41 @@ class Maybe_Show_Campaign extends Lightweight_API {
 		}
 		$campaigns = json_decode( $_REQUEST['popups'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 		$settings  = json_decode( $_REQUEST['settings'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
-		$visit     = json_decode( $_REQUEST['visit'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized, WordPress.Security.ValidatedSanitizedInput.InputNotValidated
+		$visit     = (array) json_decode( $_REQUEST['visit'] ); // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized, WordPress.Security.ValidatedSanitizedInput.InputNotValidated
 		$response  = [];
 		$client_id = $_REQUEST['cid']; // phpcs:ignore WordPress.Security.NonceVerification.Recommended, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
 
-		if ( defined( 'ENABLE_CAMPAIGN_EVENT_LOGGING' ) && ENABLE_CAMPAIGN_EVENT_LOGGING ) {
-			Segmentation_Report::api_handle_post_read(
+		if ( $visit['is_post'] && defined( 'ENABLE_CAMPAIGN_EVENT_LOGGING' ) && ENABLE_CAMPAIGN_EVENT_LOGGING ) {
+			// Update the cache.
+			$posts_read        = $this->get_client_data( $client_id )['posts_read'];
+			$already_read_post = count(
+				array_filter(
+					$posts_read,
+					function ( $post_data ) use ( $visit ) {
+						return $post_data['post_id'] == $visit['post_id'];
+					}
+				)
+			) > 0;
+
+			if ( false === $already_read_post ) {
+				$posts_read[] = [
+					'post_id'      => $visit['post_id'],
+					'category_ids' => $visit['categories'],
+				];
+				$this->save_client_data(
+					$client_id,
+					[
+						'posts_read' => $posts_read,
+					]
+				);
+			}
+
+			Segmentation_Report::log_single_visit(
 				array_merge(
 					[
 						'clientId' => $client_id,
 					],
-					(array) $visit
+					$visit
 				)
 			);
 		}
@@ -98,6 +120,8 @@ class Maybe_Show_Campaign extends Lightweight_API {
 		}
 
 		$has_newsletter_prompt = $campaign->n;
+		// Suppressing based on UTM Medium parameter in the URL.
+		$has_utm_medium_in_url = stripos( $referer_url, 'utm_medium=email' );
 
 		// Handle referer-based conditions.
 		if ( ! empty( $referer_url ) ) {
@@ -108,8 +132,6 @@ class Maybe_Show_Campaign extends Lightweight_API {
 				$campaign_data['suppress_forever'] = true;
 			}
 
-			// Suppressing based on UTM Medium parameter in the URL.
-			$has_utm_medium_in_url = stripos( $referer_url, 'utm_medium=email' );
 			if (
 				$has_utm_medium_in_url &&
 				$settings->suppress_newsletter_campaigns &&
@@ -131,6 +153,53 @@ class Maybe_Show_Campaign extends Lightweight_API {
 		) {
 			$should_display                    = false;
 			$campaign_data['suppress_forever'] = true;
+		}
+
+		$has_donated        = count( $client_data['donations'] ) > 0;
+		$has_donation_block = $campaign->d;
+
+		// Handle suppressing a donation campaign if reader is a donor and appropriate setting is active.
+		if (
+			$has_donation_block &&
+			$settings->suppress_donation_campaigns_if_donor &&
+			$has_donated
+		) {
+			$should_display                    = false;
+			$campaign_data['suppress_forever'] = true;
+		}
+
+		// Handle segmentation.
+		$campaign_segment = isset( $settings->all_segments->{$campaign->s} ) ? $settings->all_segments->{$campaign->s} : false;
+		if ( ! empty( $campaign_segment ) ) {
+			$posts_read_count = count( $client_data['posts_read'] );
+			// If coming from email, assume it's a subscriber.
+			$is_subscriber = ! empty( $client_data['email_subscriptions'] ) || $has_utm_medium_in_url;
+			$is_donor      = ! empty( $client_data['donations'] );
+			if (
+				$campaign_segment->min_posts > 0 && $campaign_segment->min_posts > $posts_read_count
+			) {
+				$should_display = false;
+			}
+			if (
+				$campaign_segment->max_posts > 0 && $campaign_segment->max_posts < $posts_read_count
+			) {
+				$should_display = false;
+			}
+			if ( $campaign_segment->is_subscribed && ! $is_subscriber ) {
+				$should_display = false;
+			}
+			if ( $campaign_segment->is_not_subscribed && $is_subscriber ) {
+				$should_display = false;
+				if ( $has_utm_medium_in_url ) { // Save suppression for this campaign.
+					$campaign_data['suppress_forever'] = true;
+				}
+			}
+			if ( $campaign_segment->is_donor && ! $is_donor ) {
+				$should_display = false;
+			}
+			if ( $campaign_segment->is_not_donor && $is_donor ) {
+				$should_display = false;
+			}
 		}
 
 		if ( ! empty( array_diff( $init_campaign_data, $campaign_data ) ) ) {