+++ wp-content/plugins/newspack-popups/includes/class-newspack-popups-model.php
@@ -166,6 +166,7 @@ final class Newspack_Popups_Model {
 					update_post_meta( $id, $key, $value );
 					break;
 				case 'utm_suppression':
+				case 'selected_segment_id':
 					update_post_meta( $id, $key, esc_attr( $value ) );
 					break;
 				default:
@@ -193,6 +194,31 @@ final class Newspack_Popups_Model {
 			'meta_key'    => 'placement',
 			// phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_value
 			'meta_value'  => 'inline',
+		];
+
+		return self::retrieve_popups_with_query( new WP_Query( $args ) );
+	}
+
+	/**
+	 * Get overlay test popups.
+	 *
+	 * @return array Overlay test popup objects.
+	 */
+	public static function retrieve_overlay_test_popups() {
+		$args = [
+			'post_type'   => Newspack_Popups::NEWSPACK_PLUGINS_CPT,
+			'post_status' => 'publish',
+			'meta_query'  => [ // phpcs:ignore WordPress.DB.SlowDBQuery.slow_db_query_meta_query
+				[
+					'key'     => 'placement',
+					'value'   => 'inline',
+					'compare' => '!=',
+				],
+				[
+					'key'   => 'frequency',
+					'value' => 'test',
+				],
+			],
 		];
 
 		return self::retrieve_popups_with_query( new WP_Query( $args ) );
@@ -245,6 +271,7 @@ final class Newspack_Popups_Model {
 				'background_color'        => filter_input( INPUT_GET, 'background_color', FILTER_SANITIZE_STRING ),
 				'display_title'           => filter_input( INPUT_GET, 'display_title', FILTER_VALIDATE_BOOLEAN ),
 				'dismiss_text'            => filter_input( INPUT_GET, 'dismiss_text', FILTER_SANITIZE_STRING ),
+				'dismiss_text_alignment'  => filter_input( INPUT_GET, 'dismiss_text_alignment', FILTER_SANITIZE_STRING ),
 				'frequency'               => filter_input( INPUT_GET, 'frequency', FILTER_SANITIZE_STRING ),
 				'overlay_color'           => filter_input( INPUT_GET, 'overlay_color', FILTER_SANITIZE_STRING ),
 				'overlay_opacity'         => filter_input( INPUT_GET, 'overlay_opacity', FILTER_SANITIZE_STRING ),
@@ -314,6 +341,7 @@ final class Newspack_Popups_Model {
 		$post_options = isset( $options ) ? $options : [
 			'background_color'        => get_post_meta( $id, 'background_color', true ),
 			'dismiss_text'            => get_post_meta( $id, 'dismiss_text', true ),
+			'dismiss_text_alignment'  => get_post_meta( $id, 'dismiss_text_alignment', true ),
 			'display_title'           => get_post_meta( $id, 'display_title', true ),
 			'frequency'               => get_post_meta( $id, 'frequency', true ),
 			'overlay_color'           => get_post_meta( $id, 'overlay_color', true ),
@@ -323,6 +351,7 @@ final class Newspack_Popups_Model {
 			'trigger_delay'           => get_post_meta( $id, 'trigger_delay', true ),
 			'trigger_scroll_progress' => get_post_meta( $id, 'trigger_scroll_progress', true ),
 			'utm_suppression'         => get_post_meta( $id, 'utm_suppression', true ),
+			'selected_segment_id'     => get_post_meta( $id, 'selected_segment_id', true ),
 		];
 
 		$popup = [
@@ -336,6 +365,7 @@ final class Newspack_Popups_Model {
 					'background_color'        => '#FFFFFF',
 					'display_title'           => false,
 					'dismiss_text'            => '',
+					'dismiss_text_alignment'  => 'center',
 					'frequency'               => 'test',
 					'overlay_color'           => '#000000',
 					'overlay_opacity'         => 30,
@@ -344,6 +374,7 @@ final class Newspack_Popups_Model {
 					'trigger_delay'           => 0,
 					'trigger_scroll_progress' => 0,
 					'utm_suppression'         => null,
+					'selected_segment_id'     => '',
 				]
 			),
 		];
@@ -381,6 +412,16 @@ final class Newspack_Popups_Model {
 	}
 
 	/**
+	 * Get the popup dismiss button alignment. Default/empty === center alignment.
+	 *
+	 * @param object $popup The popup object.
+	 * @return string|null Dismiss button alignment.
+	 */
+	protected static function get_dismiss_text_alignment( $popup ) {
+		return ! empty( $popup['options']['dismiss_text_alignment'] ) ? $popup['options']['dismiss_text_alignment'] : 'center';
+	}
+
+	/**
 	 * Get the popup delay in milliseconds.
 	 *
 	 * @param object $popup The popup object.
@@ -396,7 +437,7 @@ final class Newspack_Popups_Model {
 	 * @param object $popup The popup object.
 	 * @return boolean True if it is an inline popup.
 	 */
-	protected static function is_inline( $popup ) {
+	public static function is_inline( $popup ) {
 		return 'inline' === $popup['options']['placement'];
 	}
 
@@ -407,7 +448,20 @@ final class Newspack_Popups_Model {
 	 * @return boolean True if popup has a newsletter prompt.
 	 */
 	public static function has_newsletter_prompt( $popup ) {
-		return false !== strpos( $popup['content'], 'wp:jetpack/mailchimp' );
+		return (
+			false !== strpos( $popup['content'], 'wp:jetpack/mailchimp' ) ||
+			false !== strpos( $popup['content'], 'wp:mailchimp-for-wp/form' )
+		);
+	}
+
+	/**
+	 * Does the popup have a donation block?
+	 *
+	 * @param object $popup The popup object.
+	 * @return boolean True if popup has a donation block.
+	 */
+	public static function has_donation_block( $popup ) {
+		return false !== strpos( $popup['content'], 'wp:newspack-blocks/donate' );
 	}
 
 	/**
@@ -419,7 +473,11 @@ final class Newspack_Popups_Model {
 	 * @return string Prints the generated amp-analytics element.
 	 */
 	protected static function insert_event_tracking( $popup, $body, $element_id ) {
-		if ( Newspack_Popups::previewed_popup_id() || 'test' === $popup['options']['frequency'] ) {
+		if (
+			Newspack_Popups::previewed_popup_id() ||
+			'test' === $popup['options']['frequency'] ||
+			Newspack_Popups_Settings::is_non_interactive()
+		) {
 			return '';
 		}
 		global $wp;
@@ -428,74 +486,54 @@ final class Newspack_Popups_Model {
 
 		// Mailchimp.
 		$mailchimp_form_selector = '';
+		$email_form_field_name   = 'email';
 		if ( preg_match( '/wp-block-jetpack-mailchimp/', $body ) !== 0 ) {
 			$mailchimp_form_selector = '.wp-block-jetpack-mailchimp form';
 		}
 		if ( preg_match( '/mc4wp-form/', $body ) !== 0 ) {
 			$mailchimp_form_selector = '.mc4wp-form';
+			$email_form_field_name   = 'EMAIL';
+		}
+
+		$amp_analytics_config = [
+			'requests' => [
+				'event' => esc_url( $endpoint ),
+			],
+			'triggers' => [
+				'trackPageview' => [
+					'on'             => 'visible',
+					'request'        => 'event',
+					'visibilitySpec' => [
+						'selector'             => '#' . esc_attr( $element_id ),
+						'visiblePercentageMin' => 90,
+						'totalTimeMin'         => 500,
+						'continuousTimeMin'    => 200,
+					],
+					'extraUrlParams' => [
+						'popup_id' => esc_attr( self::canonize_popup_id( $popup['id'] ) ),
+						'cid'      => 'CLIENT_ID(' . esc_attr( Newspack_Popups_Segmentation::NEWSPACK_SEGMENTATION_CID_NAME ) . ')',
+					],
+				],
+			],
+		];
+		if ( $mailchimp_form_selector ) {
+			$amp_analytics_config['triggers']['formSubmitSuccess'] = [
+				'on'             => 'amp-form-submit-success',
+				'request'        => 'event',
+				'selector'       => '#' . esc_attr( $element_id ) . ' ' . esc_attr( $mailchimp_form_selector ),
+				'extraUrlParams' => [
+					'popup_id'            => esc_attr( self::canonize_popup_id( $popup['id'] ) ),
+					'cid'                 => 'CLIENT_ID(' . esc_attr( Newspack_Popups_Segmentation::NEWSPACK_SEGMENTATION_CID_NAME ) . ')',
+					'mailing_list_status' => 'subscribed',
+					'email'               => '$[formFields[' . esc_attr( $email_form_field_name ) . ']}',
+				],
+			];
 		}
 
 		?>
-		<?php if ( $mailchimp_form_selector ) : ?>
-			<amp-analytics>
-				<script type="application/json">
-					{
-						"requests": {
-							"event": "<?php echo esc_url( $endpoint ); ?>"
-						},
-						"triggers": {
-							"formSubmitSuccess": {
-								"on": "amp-form-submit-success",
-								"request": "event",
-								"selector": "#<?php echo esc_attr( $element_id ); ?> <?php echo esc_attr( $mailchimp_form_selector ); ?>",
-								"extraUrlParams": {
-									"popup_id": "<?php echo esc_attr( self::canonize_popup_id( $popup['id'] ) ); ?>",
-									"cid": "CLIENT_ID( <?php echo esc_attr( Newspack_Popups_Segmentation::NEWSPACK_SEGMENTATION_CID_NAME ); ?> )",
-									"mailing_list_status": "subscribed",
-									"email": "${formFields[email]}"
-								}
-							}
-						},
-						"transport": {
-							"beacon": true,
-							"xhrpost": true,
-							"useBody": true,
-							"image": false
-						}
-					}
-				</script>
-			</amp-analytics>
-		<?php endif; ?>
-
 		<amp-analytics>
 			<script type="application/json">
-				{
-					"requests": {
-						"event": "<?php echo esc_url( $endpoint ); ?>"
-					},
-					"triggers": {
-						"trackPageview": {
-							"on": "visible",
-							"request": "event",
-							"visibilitySpec": {
-								"selector": "#<?php echo esc_attr( $element_id ); ?>",
-								"visiblePercentageMin": 90,
-								"totalTimeMin": 500,
-								"continuousTimeMin": 200
-							},
-							"extraUrlParams": {
-								"popup_id": "<?php echo esc_attr( self::canonize_popup_id( $popup['id'] ) ); ?>",
-								"cid": "CLIENT_ID( <?php echo esc_attr( Newspack_Popups_Segmentation::NEWSPACK_SEGMENTATION_CID_NAME ); ?> )"
-							}
-						}
-					},
-					"transport": {
-						"beacon": true,
-						"xhrpost": true,
-						"useBody": true,
-						"image": false
-					}
-				}
+				<?php echo wp_json_encode( $amp_analytics_config ); ?>
 			</script>
 		</amp-analytics>
 		<?php
@@ -598,27 +636,6 @@ final class Newspack_Popups_Model {
 	}
 
 	/**
-	 * Add "newspack-popups-content-block" class name to a block.
-	 * This way a block rendered inside of a popup can be easily told apart.
-	 * This will only work in dynamic blocks.
-	 *
-	 * @param object $block A block.
-	 * @return object Block with className appended.
-	 */
-	public static function append_class_to_block( $block ) {
-		$class_name = 'newspack-popups-content-block';
-		if ( isset( $block['attrs']['className'] ) ) {
-			$block['attrs']['className'] = $block['attrs']['className'] . ' ' . $class_name;
-		} else {
-			$block['attrs']['className'] = $class_name;
-		}
-
-		$block['innerBlocks'] = array_map( [ __CLASS__, 'append_class_to_block' ], $block['innerBlocks'] );
-
-		return $block;
-	}
-
-	/**
 	 * Canonise popups id. The id from WP will be an integer, but AMP does not play well with that and needs a string.
 	 *
 	 * @param int $popup_id Popup id.
@@ -633,6 +650,9 @@ final class Newspack_Popups_Model {
 	 * @param object $popup Popup.
 	 */
 	public static function get_access_attrs( $popup ) {
+		if ( Newspack_Popups_Settings::is_non_interactive() ) {
+			return '';
+		}
 		if (
 			( 'test' === $popup['options']['frequency'] || Newspack_Popups::previewed_popup_id() ) &&
 			is_user_logged_in() &&
@@ -656,19 +676,20 @@ final class Newspack_Popups_Model {
 		$blocks = parse_blocks( $popup['content'] );
 		$body   = '';
 		foreach ( $blocks as $block ) {
-			$body .= render_block( self::append_class_to_block( $block ) );
+			$body .= render_block( $block );
 		}
 		do_action( 'newspack_campaigns_after_campaign_render', $popup );
 
-		$element_id           = 'lightbox' . rand(); // phpcs:ignore WordPress.WP.AlternativeFunctions.rand_rand
-		$endpoint             = self::get_reader_endpoint();
-		$display_title        = $popup['options']['display_title'];
-		$hidden_fields        = self::get_hidden_fields( $popup );
-		$dismiss_text         = self::get_dismiss_text( $popup );
-		$is_newsletter_prompt = self::has_newsletter_prompt( $popup );
-		$classes              = array( 'newspack-inline-popup' );
-		$classes[]            = ( ! empty( $popup['title'] ) && $display_title ) ? 'newspack-lightbox-has-title' : null;
-		$classes[]            = $is_newsletter_prompt ? 'newspack-newsletter-prompt-inline' : null;
+		$element_id             = 'lightbox' . rand(); // phpcs:ignore WordPress.WP.AlternativeFunctions.rand_rand
+		$endpoint               = self::get_reader_endpoint();
+		$display_title          = $popup['options']['display_title'];
+		$hidden_fields          = self::get_hidden_fields( $popup );
+		$dismiss_text           = self::get_dismiss_text( $popup );
+		$dismiss_text_alignment = self::get_dismiss_text_alignment( $popup );
+		$is_newsletter_prompt   = self::has_newsletter_prompt( $popup );
+		$classes                = array( 'newspack-inline-popup' );
+		$classes[]              = ( ! empty( $popup['title'] ) && $display_title ) ? 'newspack-lightbox-has-title' : null;
+		$classes[]              = $is_newsletter_prompt ? 'newspack-newsletter-prompt-inline' : null;
 
 		$analytics_events = self::get_analytics_events( $popup, $body, $element_id );
 		if ( ! empty( $analytics_events ) ) {
@@ -688,8 +709,8 @@ final class Newspack_Popups_Model {
 					<h1 class="newspack-popup-title"><?php echo esc_html( $popup['title'] ); ?></h1>
 				<?php endif; ?>
 						<?php echo ( $body ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
-						<?php if ( $dismiss_text ) : ?>
-					<form class="popup-not-interested-form popup-action-form"
+						<?php if ( $dismiss_text && ! Newspack_Popups_Settings::is_non_interactive() ) : ?>
+					<form class="popup-not-interested-form popup-action-form align-<?php echo esc_attr( $dismiss_text_alignment ); ?>"
 						method="POST"
 						action-xhr="<?php echo esc_url( $endpoint ); ?>"
 						target="_top">
@@ -723,21 +744,23 @@ final class Newspack_Popups_Model {
 		$blocks = parse_blocks( $popup['content'] );
 		$body   = '';
 		foreach ( $blocks as $block ) {
-			$body .= render_block( self::append_class_to_block( $block ) );
+			$body .= render_block( $block );
 		}
 		do_action( 'newspack_campaigns_after_campaign_render', $popup );
 
-		$element_id           = 'lightbox' . rand(); // phpcs:ignore WordPress.WP.AlternativeFunctions.rand_rand
-		$endpoint             = self::get_reader_endpoint();
-		$dismiss_text         = self::get_dismiss_text( $popup );
-		$display_title        = $popup['options']['display_title'];
-		$overlay_opacity      = absint( $popup['options']['overlay_opacity'] ) / 100;
-		$overlay_color        = $popup['options']['overlay_color'];
-		$hidden_fields        = self::get_hidden_fields( $popup );
-		$is_newsletter_prompt = self::has_newsletter_prompt( $popup );
-		$classes              = array( 'newspack-lightbox', 'newspack-lightbox-placement-' . $popup['options']['placement'] );
-		$classes[]            = ( ! empty( $popup['title'] ) && $display_title ) ? 'newspack-lightbox-has-title' : null;
-		$classes[]            = $is_newsletter_prompt ? 'newspack-newsletter-prompt-overlay' : null;
+		$element_id             = 'lightbox' . rand(); // phpcs:ignore WordPress.WP.AlternativeFunctions.rand_rand
+		$endpoint               = self::get_reader_endpoint();
+		$dismiss_text           = self::get_dismiss_text( $popup );
+		$dismiss_text_alignment = self::get_dismiss_text_alignment( $popup );
+		$display_title          = $popup['options']['display_title'];
+		$overlay_opacity        = absint( $popup['options']['overlay_opacity'] ) / 100;
+		$overlay_color          = $popup['options']['overlay_color'];
+		$hidden_fields          = self::get_hidden_fields( $popup );
+		$is_newsletter_prompt   = self::has_newsletter_prompt( $popup );
+		$classes                = array( 'newspack-lightbox', 'newspack-lightbox-placement-' . $popup['options']['placement'] );
+		$classes[]              = ( ! empty( $popup['title'] ) && $display_title ) ? 'newspack-lightbox-has-title' : null;
+		$classes[]              = $is_newsletter_prompt ? 'newspack-newsletter-prompt-overlay' : null;
+		$is_scroll_triggered    = 'scroll' === $popup['options']['trigger_type'];
 
 		add_filter(
 			'newspack_analytics_events',
@@ -756,7 +779,7 @@ final class Newspack_Popups_Model {
 					<?php endif; ?>
 					<?php echo ( $body ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped ?>
 					<?php if ( $dismiss_text ) : ?>
-					<form class="popup-not-interested-form popup-action-form"
+					<form class="popup-not-interested-form popup-action-form align-<?php echo esc_attr( $dismiss_text_alignment ); ?>"
 						method="POST"
 						action-xhr="<?php echo esc_url( $endpoint ); ?>"
 						target="_top">
@@ -769,7 +792,7 @@ final class Newspack_Popups_Model {
 						<button on="tap:<?php echo esc_attr( $element_id ); ?>.hide" aria-label="<?php esc_attr( $dismiss_text ); ?>" style="<?php echo esc_attr( self::container_style( $popup ) ); ?>"><?php echo esc_attr( $dismiss_text ); ?></button>
 					</form>
 					<?php endif; ?>
-					<form class="popup-dismiss-form popup-action-form"
+					<form class="popup-dismiss-form popup-action-form align-<?php echo esc_attr( $dismiss_text_alignment ); ?>"
 						method="POST"
 						action-xhr="<?php echo esc_url( $endpoint ); ?>"
 						target="_top">
@@ -788,9 +811,11 @@ final class Newspack_Popups_Model {
 				<button style="opacity: <?php echo floatval( $overlay_opacity ); ?>;background-color:<?php echo esc_attr( $overlay_color ); ?>;" class="newspack-lightbox-shim" on="tap:<?php echo esc_attr( $element_id ); ?>.hide"></button>
 			</form>
 		</amp-layout>
-		<div id="page-position-marker" style="position: absolute; top: <?php echo esc_attr( $popup['options']['trigger_scroll_progress'] ); ?>%"></div>
-		<amp-position-observer target="page-position-marker" on="enter:showAnim.start;" once layout="nodisplay"></amp-position-observer>
-		<amp-animation id="showAnim" layout="nodisplay">
+		<?php if ( $is_scroll_triggered ) : ?>
+			<div id="page-position-marker" style="position: absolute; top: <?php echo esc_attr( $popup['options']['trigger_scroll_progress'] ); ?>%"></div>
+			<amp-position-observer target="page-position-marker" on="enter:showAnim.start;" once layout="nodisplay"></amp-position-observer>
+		<?php endif; ?>
+		<amp-animation id="showAnim" layout="nodisplay" <?php echo $is_scroll_triggered ? '' : 'trigger="visibility"'; ?>>
 			<script type="application/json">
 				{
 					"duration": "125ms",