+++ wp-content/plugins/newspack-popups/includes/class-newspack-popups-segmentation.php
@@ -31,6 +31,11 @@ final class Newspack_Popups_Segmentation {
 	const NEWSPACK_SEGMENTATION_CID_LINKER_PARAM = 'ref_newspack_cid';
 
 	/**
+	 * Name of the option to store segments under.
+	 */
+	const SEGMENTS_OPTION_NAME = 'newspack_popups_segments';
+
+	/**
 	 * Main Newspack Segmentation Plugin Instance.
 	 * Ensures only one instance of Newspack Segmentation Plugin Instance is loaded or can be loaded.
 	 *
@@ -49,7 +54,6 @@ final class Newspack_Popups_Segmentation {
 	public function __construct() {
 		add_action( 'init', [ __CLASS__, 'create_database_table' ] );
 		add_action( 'wp_footer', [ __CLASS__, 'insert_amp_analytics' ], 20 );
-		add_action( 'wp_enqueue_scripts', [ __CLASS__, 'wp_enqueue_scripts' ] );
 	}
 
 	/**
@@ -63,46 +67,10 @@ final class Newspack_Popups_Segmentation {
 	 * Should tracking code be inserted?
 	 */
 	public static function is_tracking() {
-		if ( is_admin() || self::is_admin_user() ) {
+		if ( is_admin() || self::is_admin_user() || Newspack_Popups_Settings::is_non_interactive() ) {
 			return false;
 		}
 		return true;
-	}
-
-	/**
-	 * Insert amp-analytics scripts.
-	 */
-	public static function wp_enqueue_scripts() {
-		if ( ! self::is_tracking() ) {
-			return;
-		}
-
-		// Register AMP scripts explicitly for non-AMP pages.
-		if ( ! wp_script_is( 'amp-runtime', 'registered' ) ) {
-			// phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
-			wp_register_script(
-				'amp-runtime',
-				'https://cdn.ampproject.org/v0.js',
-				null,
-				null,
-				true
-			);
-		}
-		$scripts = [ 'amp-analytics' ];
-		foreach ( $scripts as $script ) {
-			if ( ! wp_script_is( $script, 'registered' ) ) {
-				$path = "https://cdn.ampproject.org/v0/{$script}-latest.js";
-				// phpcs:ignore WordPress.WP.EnqueuedResourceParameters.MissingVersion
-				wp_register_script(
-					$script,
-					$path,
-					array( 'amp-runtime' ),
-					null,
-					true
-				);
-			}
-			wp_enqueue_script( $script );
-		}
 	}
 
 	/**
@@ -142,6 +110,23 @@ final class Newspack_Popups_Segmentation {
 				],
 			],
 		];
+
+		// Handle Mailchimp URL parameters.
+		if ( isset( $_GET['mc_cid'], $_GET['mc_eid'] ) ) { // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+			$endpoint                         = self::get_client_data_endpoint();
+			$amp_analytics_config['requests'] = [
+				'event' => esc_url( $endpoint ),
+			];
+			$amp_analytics_config['triggers']['trackMailchimpData'] = [
+				'on'             => 'ini-load',
+				'request'        => 'event',
+				'extraUrlParams' => [
+					'mc_cid'    => sanitize_text_field( $_GET['mc_cid'] ), // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+					'mc_eid'    => sanitize_text_field( $_GET['mc_eid'] ), // phpcs:ignore WordPress.Security.NonceVerification.Recommended
+					'client_id' => 'CLIENT_ID(' . esc_attr( self::NEWSPACK_SEGMENTATION_CID_NAME ) . ')',
+				],
+			];
+		}
 
 		?>
 			<amp-analytics>
@@ -184,5 +169,100 @@ final class Newspack_Popups_Segmentation {
 		}
 	}
 
+	/**
+	 * Get all configured segments.
+	 */
+	public static function get_segments() {
+		return get_option( self::SEGMENTS_OPTION_NAME, [] );
+	}
+
+	/**
+	 * Create a segment.
+	 *
+	 * @param object $segment A segment.
+	 */
+	public static function create_segment( $segment ) {
+		$segments              = self::get_segments();
+		$segment['id']         = uniqid();
+		$segment['created_at'] = gmdate( 'Y-m-d' );
+		$segment['updated_at'] = gmdate( 'Y-m-d' );
+		$segments[]            = $segment;
+
+		update_option( self::SEGMENTS_OPTION_NAME, $segments );
+		return self::get_segments();
+	}
+
+	/**
+	 * Delete a segment.
+	 *
+	 * @param string $id A segment id.
+	 */
+	public static function delete_segment( $id ) {
+		$segments = array_values(
+			array_filter(
+				self::get_segments(),
+				function( $segment ) use ( $id ) {
+					return $segment['id'] !== $id;
+				}
+			)
+		);
+		update_option( self::SEGMENTS_OPTION_NAME, $segments );
+		return self::get_segments();
+	}
+
+	/**
+	 * Update a segment.
+	 *
+	 * @param object $segment A segment.
+	 */
+	public static function update_segment( $segment ) {
+		$segments              = self::get_segments();
+		$segment['updated_at'] = gmdate( 'Y-m-d' );
+		foreach ( $segments as &$_segment ) {
+			if ( $_segment['id'] === $segment['id'] ) {
+				$_segment['name']          = $segment['name'];
+				$_segment['configuration'] = $segment['configuration'];
+			}
+		}
+
+		update_option( self::SEGMENTS_OPTION_NAME, $segments );
+		return self::get_segments();
+	}
+
+	/**
+	 * Get current client's id.
+	 */
+	public static function get_client_id() {
+		return isset( $_COOKIE[ self::NEWSPACK_SEGMENTATION_CID_NAME ] ) ? esc_attr( $_COOKIE[ self::NEWSPACK_SEGMENTATION_CID_NAME ] ) : false; // phpcs:ignore WordPressVIPMinimum.Variables.RestrictedVariables.cache_constraints___COOKIE, WordPress.Security.ValidatedSanitizedInput.InputNotSanitized
+	}
+
+	/**
+	 * Get API client data endpoint.
+	 */
+	public static function get_client_data_endpoint() {
+		return plugins_url( '../api/segmentation/index.php', __FILE__ );
+	}
+
+	/**
+	 * Update client data.
+	 *
+	 * @param string $client_id Client ID.
+	 * @param string $payload Client data.
+	 */
+	public static function update_client_data( $client_id, $payload ) {
+		if ( isset( $client_id ) ) {
+			wp_safe_remote_post(
+				self::get_client_data_endpoint(),
+				[
+					'body' => array_merge(
+						[
+							'client_id' => $client_id,
+						],
+						$payload
+					),
+				]
+			);
+		}
+	}
 }
 Newspack_Popups_Segmentation::instance();